Moving Question schemas test to separate files
